requirements.txt
fastapi==0.115.0
motor==3.6.0
pydantic[email]==2.9.2
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
uvicorn==0.32.0
python-dotenv==1.0.1
pydantic-settings==2.5.2
pytest==8.3.2
httpx==0.27.2
pytest-asyncio==0.23.8
bcrypt==3.2.0
